ioclient: stopping.
